requirements.txt
apturl==0.5.2
argon2-cffi==21.1.0
asgiref==3.4.1
attrs==21.2.0
autopep8==1.5.7
awsebcli==3.20.2
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bcrypt==3.1.7
beautifulsoup4==4.10.0
bleach==4.1.0
blessed==1.19.0
blinker==1.4
botocore==1.21.65
Brlapi==0.8.2
bs4==0.0.1
cached-property==1.5.2
cement==2.8.2
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.9
chrome-gnome-shell==0.0.0
cliapp==1.20180812.1
click==8.0.1
cmdtest==0.32+git
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==3.3.2
cssselect==1.1.0
cupshelpers==1.0
dbus-python==1.2.16
debugpy==1.5.0
decorator==5.1.0
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.3
distro==1.5.0
distro-info==1.0
dj-database-url==0.5.0
Django==3.2.9
django-ckeditor==6.2.0
django-countries==7.2.1
django-embed-video==0.6
django-heroku==0.3.1
django-js-asset==1.2.2
django-smtp-ssl==1.0
dnspython==2.1.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.20
entrypoints==0.3
evdev==1.4.0
fasteners==0.14.1
filelock==3.0.12
Flask==2.0.1
flask-paginate==0.8.1
future==0.16.0
Guake==3.6.3
gunicorn==20.1.0
html5lib==1.1
httplib2==0.18.1
idna==3.3
importlib-metadata==4.0.1
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.5
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keyring==23.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lockfile==0.12.2
louis==3.18.0
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
natsort==7.1.0
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.9
nltk==3.6.3
notebook==6.4.4
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
packaging==21.0
pandocfilters==1.5.0
paramiko==2.7.2
parso==0.8.2
pathspec==0.5.9
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
pipenv==2021.5.29
platformdirs==2.3.0
ply==3.11
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.12.4
psutil==5.8.0
psycopg2==2.9.2
ptyprocess==0.7.0
PyAudio==0.2.11
pycairo==1.16.2
pycodestyle==2.7.0
pycparser==2.20
pycups==2.0.1
Pygments==2.10.0
PyGObject==3.40.1
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.12.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.0
PySocks==1.7.1
pyspellchecker==0.6.2
python-apt==2.2.1
python-dateutil==2.8.1
python-debian===0.1.39ubuntu1
python-decouple==3.5
python-dotenv==0.19.0
python-magic==0.4.20
pytube3==9.6.4
pytz==2021.3
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.3.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
regex==2021.9.30
reportlab==3.5.66
requests==2.26.0
requests-futures==1.0.0
scour==0.38.2
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
semantic-version==2.8.5
Send2Trash==1.8.0
setproctitle==1.2.1
simplejson==3.17.2
six==1.14.0
soupsieve==2.2.1
sqlparse==0.4.2
stem==1.8.0
stripe==2.63.0
systemd-python==234
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
texttable==1.6.4
TLPUI==1.4.0a0
toml==0.10.2
tornado==6.1
torrequest==0.1.0
tqdm==4.62.3
traitlets==5.1.0
ttystatus==0.38
typing-extensions==3.10.0.2
ubuntu-advantage-tools==27.4
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.7
usb-creator==0.3.7
variety==0.8.5
virtualenv==20.8.1
virtualenv-clone==0.5.7
wadllib==1.3.5
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.59.0
websockets==8.1
Werkzeug==2.0.1
whitenoise==5.3.0
widgetsnbextension==3.5.1
wxPython==4.0.7
xdg==5
xkit==0.0.0
youtube-dl==2021.6.6
zipp==1.0.0
